Added exceptions.

+++ src/main/java/com/usertest/service/userservice/UserServiceImpl.java
@@ -4,6 +4,8 @@ import com.usertest.dto.AddressDto;
 import com.usertest.dto.UserDto;
 import com.usertest.entity.UserEntity;
 import com.usertest.entity.UserWithNumberEntity;
+import com.usertest.exception.NotFoundNumberException;
+import com.usertest.exception.NotFoundUserException;
 import com.usertest.mapper.UserMapper;
 import com.usertest.repository.NumberRepository;
 import com.usertest.repository.UserRepository;
@@ -30,7 +32,8 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto getUserById(long id) {
-        UserEntity userEntity = userRepository.getUserById(id);
+        UserEntity userEntity = userRepository.getUserById(id).orElseThrow(() ->
+                new NotFoundUserException("There is no user with ID = " + id + " in database."));
         var responseAddressDto = restService.getAddressById(userEntity.getAddressId());
         List<String> numberEntitiesList = numberRepository.getNumbersByUserId(id);
         return userMapper.toUserDto(userEntity, numberEntitiesList, responseAddressDto.getData());
@@ -69,19 +72,41 @@ public class UserServiceImpl implements UserService {
     public UserDto updateUser(long userId, UserDto userDto) {
         var addressDto = new AddressDto();
         addressDto.setAddress(userDto.getAddress());
+        userRepository.getUserById(userId).orElseThrow(() ->
+                new NotFoundUserException("There is no user with ID = " + userId + " in database."));
         var addressResult = restService.findOrInsertAddress(addressDto);
+
         UserEntity userEntity = userRepository.updateUser(userMapper.toUserEntity(userDto, addressResult.getData().getId()));
-        numberRepository.deleteNumbersByUserId(userId);
-        numberRepository.saveNumbersList(userDto.getNumbers(), userId);
+        List<String> existNumbers = numberRepository.getNumbersByUserId(userId);
+        if(existNumbers != null && !existNumbers.isEmpty()) {
+            int deleteNumbersResult = numberRepository.deleteNumbersByUserId(userId);
+            if (deleteNumbersResult == 0) {
+                throw new NotFoundNumberException("There is no numbers with user ID = " + userId + " in database.");
+            }
+        }
+        if (userDto.getNumbers() != null && !userDto.getNumbers().isEmpty()) {
+            numberRepository.saveNumbersList(userDto.getNumbers(), userId);
+        }
+
         return userMapper.toUserDto(userEntity, userDto.getNumbers(), addressResult.getData());
     }
 
     @Transactional
     @Override
-    public int deleteUserById(long id) {
-        var userDto = userRepository.getUserById(id);
-        int result = userRepository.deleteUserById(id);
-        numberRepository.deleteNumbersByUserId(id);
+    public int deleteUserById(long userId) {
+        var userDto = userRepository.getUserById(userId).orElseThrow(() ->
+                new NotFoundUserException("There is no user with ID = " + userId + " in database."));
+        int result = userRepository.deleteUserById(userId);
+        if (result == 0) {
+            throw new NotFoundUserException("There is no user with ID = " + userId + " in database.");
+        }
+        List<String> existNumbers = numberRepository.getNumbersByUserId(userId);
+        if (existNumbers != null && !existNumbers.isEmpty()) {
+            int deleteNumbersResult = numberRepository.deleteNumbersByUserId(userId);
+            if (deleteNumbersResult == 0) {
+                throw new NotFoundNumberException("There is no numbers with user ID = " + userId + " in database.");
+            }
+        }
         restService.deleteAddressById(userDto.getAddressId());
         return result;
     }
